backtesting-platform/requirements.txt
yfinance
ta
pandas
streamlit
sqlite3
apache-airflow
